+++ glue-examples/src/main/java/com/example/CustomCryptoSettings.java
@@ -11,8 +11,10 @@ import org.bouncycastle.operator.InputDecryptorProvider;
 import org.bouncycastle.operator.OperatorCreationException;
 import org.bouncycastle.pkcs.PKCS8EncryptedPrivateKeyInfo;
 import org.bouncycastle.pkcs.PKCSException;
+import org.somda.sdc.dpws.crypto.CachingCryptoSettings;
 import org.somda.sdc.dpws.crypto.CryptoSettings;
 
+import javax.net.ssl.SSLContext;
 import java.io.BufferedReader;
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
@@ -33,7 +35,7 @@ import java.security.cert.CertificateFactory;
 import java.util.Objects;
 import java.util.Optional;
 
-public class CustomCryptoSettings implements CryptoSettings {
+public class CustomCryptoSettings implements CachingCryptoSettings {
     private static final Logger LOG = LogManager.getLogger(CustomCryptoSettings.class);
 
     private static final String DEFAULT_KEYSTORE = "crypto/sdcparticipant.jks";
@@ -45,6 +47,8 @@ public class CustomCryptoSettings implements CryptoSettings {
     private byte[] trustStore = null;
     private String keyStorePassword = null;
     private String trustStorePassword = null;
+
+    private Optional<SSLContext> cachedContext = Optional.empty();
 
     public CustomCryptoSettings(
             byte[] keyStore,
@@ -203,5 +207,15 @@ public class CustomCryptoSettings implements CryptoSettings {
     @Override
     public String getTrustStorePassword() {
         return Objects.requireNonNullElse(trustStorePassword, DEFAULT_TRUSTSTORE_PASSWORD);
+    }
+
+    @Override
+    public Optional<SSLContext> getSslContext() {
+        return cachedContext;
+    }
+
+    @Override
+    public void setSslContext(final SSLContext sslContext) {
+        cachedContext = Optional.of(sslContext);
     }
 }